fastapi/requirements.txt
uwsgi
fastapi
uvicorn
jinja2
python-multipart
llama-index
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
azure-cognitiveservices-speech
ffmpeg-python
pydub
moviepy